+++ R/paletteer_d.R
@@ -1,38 +1,35 @@
 #' Get discrete palette by package and name
 #'
 #' Available package/palette combinations are available in the data.frame
-#' \code{\link[paletteer]{palettes_d_names}}. Both `package` and `palette`
-#' can be supplied as symbols or strings.
+#' \code{\link[paletteer]{palettes_d_names}}.
 #'
-#' @param package Name of package from which the palette is desired as string
-#' or symbol.
-#' @param palette Name of palette as string or symbol.
+#' @param palette Name of palette as a string. Must be on the form
+#' packagename::palettename.
 #' @param n Number of colors desired. If omitted, returns complete palette.
 #' @param direction Either `1` or `-1`. If `-1` the palette will be reversed.
 #' @param type Either "discrete" or "continuous". Colors are interpolated if
 #'   "continuous" is picked. Defaults to "discrete".
 #' @return A vector of colors.
 #' @examples
-#' paletteer_d("nord", "frost")
-#' paletteer_d("wesanderson", "Royal1", 3)
-#' paletteer_d("Redmonder", "dPBIPuOr", 14, type = "continuous")
+#' paletteer_d("nord::frost")
+#' paletteer_d("wesanderson::Royal1", 3)
+#' paletteer_d("Redmonder::dPBIPuOr", 14, type = "continuous")
 #' @export
-paletteer_d <- function (package, palette, n, direction = 1,
+paletteer_d <- function (palette, n, direction = 1,
                          type = c("discrete", "continuous")) {
 
   if (abs(direction) != 1) {
     stop("direction must be 1 or -1")
   }
 
-  package <- rlang::quo_name(rlang::enquo(package))
-  palette <- rlang::quo_name(rlang::enquo(palette))
-
-  package <- match.arg(package, unique(paletteer::palettes_d_names$package))
-
   type <- match.arg(type)
-  pal <- paletteer::palettes_d[[c(package, palette)]]
+
+  palette <- unlist(strsplit(palette, "::"))
+
+  pal <- paletteer::palettes_d[[palette]]
   if (is.null(pal))
-    stop("Palette not found. Make sure the palette name are spelled correct.")
+    stop('Palette not found. Make sure both package and palette ',
+         'name are spelled correct in the format "package::palette"')
   if (missing(n)) {
     n <- length(pal)
   }